test: a user can not delete himself

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
-use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
@@ -15,6 +15,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        //
         return view('admin.users.index');
     }
 
@@ -23,10 +24,10 @@ class UserController extends Controller
      */
     public function create()
     {
-        $permissions = Permission::all();
-        $roles = Role::all();
+        //
+        $roles=Role::all();
+        return view('admin.users.create', compact('roles'));
 
-        return view('admin.users.create', compact('permissions', 'roles'));
     }
 
     /**
@@ -34,32 +35,26 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-       
-    $data = $request->validate([
-    'name' => 'required|string|max:255',
-    'email' => 'required|string|email|max:255|unique:users',
-    'password' => 'required|string|min:8|confirmed',
-    'id_number' => 'required|string|min:5|max:20|regex:/^[A-Za-z0-9]+$/|unique:users',
-    'phone' => 'required|digits_between:7,15',
-    'address' => 'required|string|max:255',
-    'role_id' => 'required|exists:roles,id',
-    ]); 
-    
-    $user =User::create($data);
+        //
+        $data = $request->validate([
+            'name' => 'required|string|max:255',
+            'email' => 'required|string|email|max:255|unique:users,email',
+            'id_number' => 'required|string|min:5|max:20|regex:/^[A-Za-z0-9]+$/|unique:users,id_number',
+            'address' => 'required|max:255',
+            'phone' => 'required|digits_between:7,15',
+            'password' => 'required|string|min:8|confirmed',
+            'role_id' => 'required|exists:roles,id'
+        ]);
+        $user = User::create($data);
+        $user->roles()->attach($data['role_id']);
 
-    $user->roles()->attach($data['role_id']);
-
-    
-    session()->flash('swal', [
-    'icon' => 'success',
-    'title' => 'Usuario creado correctamente',
-    'text' => 'El usuario ha sido creado correctamente'
-]);
-
-return redirect()->route('admin.users.index')
-    ->with('success', 'User created successfully.');
-
-
+        //alerta de funcionamiento
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => 'Usuario creado correctamente',
+            'text' => 'El usuario ha sido creado corrretamente',
+        ]);
+        return redirect(route('admin.users.index'))->with('success', 'User created successfully');
     }
 
     /**
@@ -75,10 +70,9 @@ return redirect()->route('admin.users.index')
      */
     public function edit(User $user)
     {
-        $permissions = Permission::all();
-        $roles = Role::all();
-
-        return view('admin.users.edit', compact('user', 'permissions', 'roles'));
+        $roles=Role::all();
+        //
+        return view('admin.users.edit', compact('user', 'roles'));
     }
 
     /**
@@ -86,36 +80,32 @@ return redirect()->route('admin.users.index')
      */
     public function update(Request $request, User $user)
     {
-    $data = $request->validate([
-    'name' => 'required|string|max:255',
-    'email' => 'required|string|email|max:255|unique:users,email,'. $user->id,
-    'id_number' => 'required|string|min:5|max:20|regex:/^[A-Za-z0-9]+$/|unique:users,id_number,'. $user->id,
-    'phone' => 'required|digits_between:7,15',
-    'address' => 'required|string|max:255',
-    'role_id' => 'required|exists:roles,id',
-    ]); 
-    
-    $user->update($data);
+        //
+        $data = $request->validate([
+            'name' => 'required|string|max:255, name,',
+            'email' => 'required|string|email|max:255|unique:users,email,'.$user->id,
+            'id_number' => 'required|string|min:5|max:20|regex:/^[A-Za-z0-9]+$/|unique:users,id_number,'.$user->id,
+            'address' => 'required|max:255',
+            'phone' => 'required|digits_between:7,15',
+            'role_id' => 'required|exists:roles,id'
+        ]);
+        $user->update($data);
 
-    
-    // Si el usuario quiere editar la contraseña
-    if ($request->filled('password')) {
-    $user->password = bcrypt($request->password);
-    $user->save();
-    }
+        //guardar la contrasena si desea actualizarse
+        if($request->filled('password')){
+            $user->password = bcrypt($request->password);
+            $user->save();
+        }
 
+        $user->roles()->sync($data['role_id']);
 
-
-    $user->roles()->sync($data['role_id']);
-
-    
-    session()->flash('swal', [
-    'icon' => 'success',
-    'title' => 'Usuario actualizado',
-    'text' => 'El usuario ha sido actualizado correctamente'
-]);
-
-return redirect()->route('admin.users.edit', $user->id);
+        //alerta de funcionamiento
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => 'Usuario actualizado correctamente',
+            'text' => 'El usuario ha sido actualizado corrretamente',
+        ]);
+        return redirect() -> route('admin.users.edit', $user->id);
     }
 
     /**
@@ -123,12 +113,30 @@ return redirect()->route('admin.users.edit', $user->id);
      */
     public function destroy(User $user)
     {
-        // Eliminar roles asociados a un usuario 
-        $user->roles()->detach(); 
-        // Eliminar usuario 
-        $user->delete(); session()->flash('swal', [ 'icon' => 'success', 
-        'title' => 'Usuario eliminado', 'text' => 'El usuario ha sido eliminado correctamente' ]); 
-        return redirect(route('admin.users.index'));
+        //
+
+        //evitar que el usuario logueado se borre a si mismo
+        if($user->id == Auth::user()->id){
+            session()->flash('swal',[
+                'icon' => 'error',
+                'title' => 'No puedes eliminarte a ti mismo',
+                'text' => 'No puedes eliminar tu propio usuario',
+            ]);
+            abort(403, 'No puedes eliminar tu propio usuario');
+            return redirect(route('admin.users.index'));   
+        }
+
+        //elimina los roles asociados al usuario antes de eliminarlo
+        $user->roles()->detach();
+        //elimina al usurio de la bd
+        $user->delete();
+
+        session()->flash('swal',[
+            'title' => 'Usuario eliminado correctamente',
+            'icon' => 'success',
+            'text' =>  'El usuario ha sido eliminado corrretamente',
+        ]);
+        return view('admin.users.index');
+
     }
 }
-
